# qlife-latest/qlife-server/openapi.yaml
openapi: 3.0.3
info:
  title: QLife API
  description: |
    REST API for the QLife mental-health platform. Supports three primary roles:
    - **USER** – a registered end-user seeking mental-health support
    - **PROFESSIONAL** – a verified mental-health professional
    - **ADMIN** – a platform administrator

    All protected endpoints require a Bearer JWT obtained from the Cognito-backed
    authentication flow. The token is passed in the `Authorization` header as
    `Bearer <token>`.

    All successful responses wrap their payload in a top-level `data` key.
    Error responses use `{ "error": { "code": "...", "message": "..." } }`.
  version: 1.0.0
  contact:
    name: QLife Engineering
servers:
  - url: https://api.qlife.app/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /me:
    get:
      tags: [Auth]
      operationId: getMyAccount
      summary: Get current account
      description: |
        Returns the authenticated account's identity record including `id`,
        `email`, `role`, `status`, `cognitoSub`, and `authProvider`.
        Available to any authenticated role.
      responses:
        '200':
          description: Account record
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      account:
                        $ref: '#/components/schemas/Account'
              example:
                data:
                  account:
                    id: "a1b2c3d4-0000-0000-0000-000000000001"
                    email: "user@example.com"
                    role: "USER"
                    status: "ACTIVE"
                    cognitoSub: "us-east-1:xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                    authProvider: "COGNITO"
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---------- Users ---------------------------------------------------------
...
  /professionals/register:
    post:
      tags: [Professionals]
      operationId: registerProfessional
      summary: Register as a professional
      description: |
        Converts the authenticated USER account into a PROFESSIONAL account
        (step 1 of the onboarding flow). The account role changes to
        PROFESSIONAL after this call. Requires role **USER**.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterProfessionalRequest'
            example:
              fullName: "Dr. Fatema Khanam"
              professionType: "PSYCHIATRIST"
              gender: "FEMALE"
              designation: "MBBS, MD (Psychiatry)"
              phone: "01811000001"
      responses:
        '200':
          description: Professional registration result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    description: Registration details (varies by implementation)
              example:
                data:
                  professionalProfileId: "prof-0000-0000-0000-000000000001"
                  status: "PENDING_VERIFICATION"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /professionals:
    get:
      tags: [Professionals]
      operationId: listProfessionals
      summary: Browse professional directory
      description: |
        Returns a paginated list of verified, visible professionals for the
        user-side directory. Requires role **USER**.
      parameters:
        - $ref: '#/components/parameters/PageQuery'
      responses:
        '200':
          description: Professional directory
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      professionals:
                        type: array
                        items:
                          $ref: '#/components/schemas/Professional'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
              example:
                data:
                  professionals:
                    - id: "prof-0000-0000-0000-000000000001"
                      fullName: "Dr. Fatema Khanam"
                      professionType: "PSYCHIATRIST"
                      gender: "FEMALE"
                      designation: "MBBS, MD (Psychiatry)"
                      workplace: "Dhaka Medical College"
                      yearsOfExperience: 10
                      bio: "Specialist in anxiety and mood disorders."
                      isVisible: true
                      acceptingNewClients: true
                  pagination:
                    page: 1
                    pageSize: 20
                    total: 42
                    hasMore: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /professionals/me:
    get:
      tags: [Professionals]
      operationId: getMyProfessionalProfile
      summary: Get own professional profile
      description: |
        Returns the full professional profile for the authenticated
        PROFESSIONAL account. Requires role **PROFESSIONAL**.
      responses:
        '200':
          description: Professional profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      professional:
                        $ref: '#/components/schemas/ProfessionalProfile'
              example:
                data:
                  professional:
                    id: "prof-0000-0000-0000-000000000001"
                    fullName: "Dr. Fatema Khanam"
                    professionType: "PSYCHIATRIST"
                    gender: "FEMALE"
                    designation: "MBBS, MD (Psychiatry)"
                    workplace: "Dhaka Medical College"
                    yearsOfExperience: 10
                    bio: "Specialist in anxiety and mood disorders."
                    phone: "01811000001"
                    isVisible: true
                    acceptingNewClients: true
                    isOnboardingComplete: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      tags: [Professionals]
      operationId: updateMyProfessionalProfile
      summary: Update own professional profile
      description: |
        Partially updates the authenticated PROFESSIONAL's profile. All fields
        are optional. Requires role **PROFESSIONAL**.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfessionalRequest'
            example:
              bio: "Updated bio focusing on CBT and trauma therapy."
              yearsOfExperience: 12
              acceptingNewClients: false
              isVisible: true
      responses:
        '200':
          description: Updated professional profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      professional:
                        $ref: '#/components/schemas/ProfessionalProfile'
              example:
                data:
                  professional:
                    id: "prof-0000-0000-0000-000000000001"
                    fullName: "Dr. Fatema Khanam"
                    professionType: "PSYCHIATRIST"
                    gender: "FEMALE"
                    designation: "MBBS, MD (Psychiatry)"
                    workplace: "Dhaka Medical College"
                    yearsOfExperience: 12
                    bio: "Updated bio focusing on CBT and trauma therapy."
                    phone: "01811000001"
                    isVisible: true
                    acceptingNewClients: false
                    isOnboardingComplete: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /professionals/verifications/{id}/review:
    post:
      tags: [Professionals]
      operationId: reviewVerification
      summary: Review a professional verification (Admin)
      description: |
        Allows an ADMIN to approve, reject, or revoke a professional
        verification submission. Requires role **ADMIN**.
      parameters:
        - name: id
          in: path
          required: true
          description: Verification record UUID
          schema:
            type: string
            format: uuid
            example: "v1b2c3d4-0000-0000-0000-000000000001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewVerificationRequest'
            example:
              decision: "APPROVED"
              decisionNote: "Documents verified successfully."
      responses:
        '200':
          description: Updated verification record
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      verification:
                        $ref: '#/components/schemas/Verification'
              example:
                data:
                  verification:
                    id: "v1b2c3d4-0000-0000-0000-000000000001"
                    decision: "APPROVED"
                    decisionNote: "Documents verified successfully."
                    decidedAt: "2025-06-01T10:00:00.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------- Appointments --------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by Amazon Cognito. Pass as Authorization: Bearer <token>.
  parameters:
    PageQuery:
      name: page
      in: query
      required: false
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "professionalProfileId must be a UUID"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Unauthorized"
    Forbidden:
      description: Authenticated account lacks the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "FORBIDDEN"
              message: "Forbidden"
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error description
              example: "professionalProfileId must be a UUID"
    Pagination:
      type: object
      description: Pagination metadata included in list responses
      required: [page, pageSize, total, hasMore]
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        hasMore:
          type: boolean
          example: true
...
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-0000-0000-0000-000000000001"
        email:
          type: string
          format: email
          example: "user@example.com"
        role:
          type: string
          enum: [USER, PROFESSIONAL, ADMIN]
          example: "USER"
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, DELETED]
          example: "ACTIVE"
        cognitoSub:
          type: string
          example: "us-east-1:xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
        authProvider:
          type: string
          example: "COGNITO"

    # ---- User profile -------------------------------------------------------
...
    Professional:
      type: object
      description: Summary professional record for directory listings
      properties:
        id:
          type: string
          format: uuid
          example: "prof-0000-0000-0000-000000000001"
        fullName:
          type: string
          example: "Dr. Fatema Khanam"
        professionType:
          type: string
          enum: [PSYCHIATRIST, PSYCHOLOGIST, COUNSELOR, THERAPIST, SOCIAL_WORKER, OTHER]
          example: "PSYCHIATRIST"
        gender:
          type: string
          nullable: true
          enum: [MALE, FEMALE, OTHER, PREFER_NOT_TO_SAY]
          example: "FEMALE"
        designation:
          type: string
          nullable: true
          example: "MBBS, MD (Psychiatry)"
        workplace:
          type: string
          nullable: true
          example: "Dhaka Medical College"
        yearsOfExperience:
          type: integer
          nullable: true
          example: 10
        bio:
          type: string
          nullable: true
          example: "Specialist in anxiety and mood disorders."
        isVisible:
          type: boolean
          example: true
        acceptingNewClients:
          type: boolean
          example: true
    ProfessionalProfile:
      allOf:
        - $ref: '#/components/schemas/Professional'
        - type: object
          properties:
            phone:
              type: string
              nullable: true
              example: "01811000001"
            isOnboardingComplete:
              type: boolean
              example: true
    RegisterProfessionalRequest:
      type: object
      required: [fullName, professionType]
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 120
          example: "Dr. Fatema Khanam"
        professionType:
          type: string
          enum: [PSYCHIATRIST, PSYCHOLOGIST, COUNSELOR, THERAPIST, SOCIAL_WORKER, OTHER]
          example: "PSYCHIATRIST"
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER, PREFER_NOT_TO_SAY]
          example: "FEMALE"
        designation:
          type: string
          maxLength: 80
          example: "MBBS, MD (Psychiatry)"
        phone:
          type: string
          maxLength: 32
          example: "01811000001"
    UpdateProfessionalRequest:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 120
          example: "Dr. Fatema Khanam"
        professionType:
          type: string
          enum: [PSYCHIATRIST, PSYCHOLOGIST, COUNSELOR, THERAPIST, SOCIAL_WORKER, OTHER]
          example: "PSYCHIATRIST"
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER, PREFER_NOT_TO_SAY]
          example: "FEMALE"
        designation:
          type: string
          maxLength: 80
          example: "MBBS, MD (Psychiatry)"
        workplace:
          type: string
          maxLength: 200
          example: "Dhaka Medical College"
        yearsOfExperience:
          type: integer
          minimum: 0
          example: 12
        bio:
          type: string
          maxLength: 2000
          example: "Updated bio focusing on CBT and trauma therapy."
        phone:
          type: string
          maxLength: 32
          example: "01811000001"
        isVisible:
          type: boolean
          example: true
        acceptingNewClients:
          type: boolean
          example: false
        isOnboardingComplete:
          type: boolean
          example: true
    Verification:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "v1b2c3d4-0000-0000-0000-000000000001"
        decision:
          type: string
          enum: [APPROVED, REJECTED, REVOKED]
          example: "APPROVED"
        decisionNote:
          type: string
          nullable: true
          example: "Documents verified successfully."
        decidedAt:
          type: string
          format: date-time
          example: "2025-06-01T10:00:00.000Z"
    ReviewVerificationRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [APPROVED, REJECTED, REVOKED]
          example: "APPROVED"
        decisionNote:
          type: string
          maxLength: 2000
          example: "Documents verified successfully."

    # ---- Appointments ------------------------------------------------------
...